rename: 기존 ExpenditureListResponseDto 내부 클래스 별도의 외부 클래스로 분리
- 다른 API에서의 재활용성을 위해

[SW-17]
