+++ src/main/java/com/wanted/safewallet/domain/expenditure/business/mapper/ExpenditureMapper.java
@@ -9,10 +9,10 @@ import com.wanted.safewallet.domain.expenditure.persistence.entity.Expenditure;
 import com.wanted.safewallet.domain.expenditure.web.dto.request.ExpenditureCreateRequestDto;
 import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureCreateResponseDto;
 import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureDetailsResponseDto;
+import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureListByDateResponseDto;
+import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureListByDateResponseDto.ExpenditureResponseDto;
 import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureListResponseDto;
-import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureListResponseDto.ExpenditureListByDateResponseDto;
-import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureListResponseDto.ExpenditureResponseDto;
-import com.wanted.safewallet.domain.expenditure.web.dto.response.ExpenditureListResponseDto.TotalAmountByCategoryResponseDto;
+import com.wanted.safewallet.domain.expenditure.web.dto.response.TotalAmountByCategoryResponseDto;
 import com.wanted.safewallet.domain.user.persistence.entity.User;
 import com.wanted.safewallet.global.dto.response.PageResponse;
 import java.time.LocalDate;
