+++ src/main/java/com/wanted/safewallet/domain/expenditure/web/dto/response/ExpenditureListResponseDto.java
@@ -1,8 +1,6 @@
 package com.wanted.safewallet.domain.expenditure.web.dto.response;
 
-import com.wanted.safewallet.domain.category.persistence.entity.CategoryType;
 import com.wanted.safewallet.global.dto.response.PageResponse;
-import java.time.LocalDate;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -20,42 +18,4 @@ public class ExpenditureListResponseDto {
     private List<ExpenditureListByDateResponseDto> expenditureListByDate;
 
     private PageResponse paging;
-
-    @Getter
-    @Builder
-    @AllArgsConstructor
-    public static class TotalAmountByCategoryResponseDto {
-
-        private Long categoryId;
-
-        private CategoryType type;
-
-        private Long totalAmount;
-    }
-
-    @Getter
-    @Builder
-    @AllArgsConstructor
-    public static class ExpenditureListByDateResponseDto {
-
-        private LocalDate expenditureDate;
-
-        private List<ExpenditureResponseDto> expenditureList;
-    }
-
-    @Getter
-    @Builder
-    @AllArgsConstructor
-    public static class ExpenditureResponseDto {
-
-        private Long expenditureId;
-
-        private Long amount;
-
-        private Long categoryId;
-
-        private CategoryType type;
-
-        private String note;
-    }
 }
