+++ src/main/java/com/wanted/safewallet/domain/expenditure/web/dto/response/TotalAmountByCategoryResponseDto.java
@@ -1,0 +1,18 @@
+package com.wanted.safewallet.domain.expenditure.web.dto.response;
+
+import com.wanted.safewallet.domain.category.persistence.entity.CategoryType;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class TotalAmountByCategoryResponseDto {
+
+    private Long categoryId;
+
+    private CategoryType type;
+
+    private Long totalAmount;
+}
